requirements.txt
google-cloud-bigquery==3.4.0
python-dotenv==0.19.0
google-auth==2.14.1
db-dtypes==1.0.4
google-cloud-storage==2.6.0
pandas==1.5.1
scikit-learn==1.1.2
joblib==1.2.0
matplotlib==3.6.2
seaborn==0.12.1
tensorflow==2.11.0
tensorflow_federated==0.41.0
azureml-core==1.48.0
azureml-sdk==1.48.0
waitress==2.1.2
flask==2.2.2
numpy==1.22.0
google_api_python_client==2.69.0
google-cloud-compute==1.7.0
Flask-Cors==3.0.10
